Added new chart

+++ src/pages/api/delete-transactions.ts
@@ -3,29 +3,20 @@ import type { APIRoute } from "astro";
 import { createServerClient, parseCookieHeader } from "@supabase/ssr";
 
 export const DELETE: APIRoute = async ({ request, cookies }) => {
-  const supabase = createServerClient(
-    import.meta.env.SUPABASE_URL,
-    import.meta.env.SUPABASE_KEY,
-    {
-      cookies: {
-        getAll() {
-          return parseCookieHeader(request.headers.get("Cookie") ?? "");
-        },
-        setAll(cookiesToSet) {
-          cookiesToSet.forEach(({ name, value, options }) =>
-            cookies.set(name, value, options)
-          );
-        },
+  const supabase = createServerClient(import.meta.env.SUPABASE_URL, import.meta.env.SUPABASE_KEY, {
+    cookies: {
+      getAll() {
+        return parseCookieHeader(request.headers.get("Cookie") ?? "");
       },
-    }
-  );
+      setAll(cookiesToSet) {
+        cookiesToSet.forEach(({ name, value, options }) => cookies.set(name, value, options));
+      },
+    },
+  });
 
   const userId = (await supabase.auth.getUser()).data.user?.id;
 
-  const { error } = await supabase
-    .from("transactions")
-    .delete()
-    .eq("user_id", userId);
+  const { error } = await supabase.from("transactions").delete().eq("user_id", userId);
 
   if (error) {
     return new Response(`Error deleting transactions`, {
